refactor db methods
