--- internal/db/mongodb/ops.go
@@ -1,7 +1,0 @@
-package mongodb
-
-import "context"
-
-func (m *MongoClient) AddPet(ctx context.Context) {
-
-}
